Changed written type name convention. Improved cast management.

+++ test/literals.c
@@ -5,7 +5,7 @@ int main()
 	unsigned int t = 51;
 	float f = 4.56f;
 	double d = 21.695;
-	double dd = (double)f;
+	double dd = f;
 	char c = '\t';
 	char* str = "Hello";
 	return 0;
